refactor(photon): split engine, projection, and zod wire schemas out of index.ts
