+++ photon/package.json
@@ -37,7 +37,9 @@
 		"access": "public"
 	},
 	"dependencies": {
+		"@hono/zod-openapi": "^1.4.0",
 		"@mailwoman/annotations": "workspace:*",
+		"@mailwoman/api-kit": "workspace:*",
 		"@mailwoman/ban": "workspace:*",
 		"@mailwoman/codex": "workspace:*",
 		"@mailwoman/neural": "workspace:*",
@@ -45,6 +47,8 @@
 		"@mailwoman/resolver": "workspace:*",
 		"@mailwoman/resolver-wof-sqlite": "workspace:*",
 		"express": "^5.2.1",
-		"mailwoman": "workspace:*"
+		"hono": "^4.12.29",
+		"mailwoman": "workspace:*",
+		"zod": "^4.4.3"
 	}
 }
